feat: autopolyfills
