+++ src/index.cjs
@@ -2,7 +2,7 @@
 
 const muif = module.exports = require('./exports/__index.cjs');
 
-muif.lib.es.Object.defineProperty(muif.lib.es.global, 'muif', {
+Object.defineProperty(globalThis, 'muif', {
   get: () => muif,
   set: () => false,
   configurable: false,
