+++ src/lib/ns.cjs
@@ -1,7 +1,7 @@
 'use strict';
 
 // NOTE: This module is used by other modules to make the namespace system work,
-//       DO NOT use any of the lib files we make in here besides "lnullj" and "es/globalThis".
+//       DO NOT use any of the lib files we make in here besides "lnullj".
 
 /**
  * @module muiflib/ns
@@ -9,9 +9,6 @@
 const ns = {};
 
 const lnullj = require('./lnullj.cjs');
-const global = require('./es/globalThis.cjs');
-
-/** @import T_es from './__es.cjs' */
 
 /**
  * @typedef {!(object & { exports: unknown })} INamespaceHandle_module
@@ -33,18 +30,10 @@ ns.create = /** @param {INamespaceHandle_module} modul3 @returns {INamespaceHand
       modul3.__locked = true;
       modul3.exports = lnullj.lock(export$);
     },
-    global: global, // Re-export of the global scope for ease of use.
+    global: globalThis, // Re-export of the global scope for ease of use.
 
     // Re-export some basics to make life easier.
     lnullj: lnullj,
-    // Lazilly loaded to prevent circular dependancies.
-    /**
-     * @type {typeof T_es}
-     * @returns {typeof T_es}
-     */
-    get es() {
-      return require('./__es.cjs');
-    },
   };
 };
