+++ src/modules/beans/implementations/immutable.cjs
@@ -3,7 +3,7 @@
  * @module muif/beans/immutable
  */
 
-const {lockns, es: { console }} = require('../../../lib/ns.cjs').create(module, exports);
+const {lockns} = require('../../../lib/ns.cjs').create(module, exports);
 
 const {InstanceBean, BeanSecret, constructInstanceBean} = require('../instance.cjs');
