+++ src/modules/beans/implementations/readonly.cjs
@@ -3,7 +3,7 @@
  * @module muif/beans/readonly
  */
 
-const {lockns, es: { console }} = require('../../../lib/ns.cjs').create(module, exports);
+const {lockns} = require('../../../lib/ns.cjs').create(module, exports);
 
 const {InstanceBean, BeanSecret, constructInstanceBean} = require('../instance.cjs');
