+++ src/modules/beans/virtual.cjs
@@ -3,7 +3,7 @@
  * @module muif/beans/virtual
  */
 
-const {lockns, es: { console }} = require('../../lib/ns.cjs').create(module, exports);
+const {lockns} = require('../../lib/ns.cjs').create(module, exports);
 
 /**
  * Super secret internal bean configuration key.
