+++ test.cjs
@@ -1,1 +1,2 @@
+const muif = require('.');
 const closure = require('.').lib.closure();
